Move upload utils to own folder
