+++ src/__test__/utils/upload/processUpload.test.js
@@ -5,14 +5,14 @@ import uuid from 'uuid';
 
 import { Storage } from 'aws-amplify';
 import fetchMock, { enableFetchMocks } from 'jest-fetch-mock';
-import { SAMPLES_FILE_UPDATE } from '../../redux/actionTypes/samples';
+import { SAMPLES_FILE_UPDATE } from '../../../redux/actionTypes/samples';
 
-import initialSampleState, { sampleTemplate } from '../../redux/reducers/samples/initialState';
-import initialProjectState, { projectTemplate } from '../../redux/reducers/projects/initialState';
-import initialExperimentState, { experimentTemplate } from '../../redux/reducers/experiments/initialState';
+import initialSampleState, { sampleTemplate } from '../../../redux/reducers/samples/initialState';
+import initialProjectState, { projectTemplate } from '../../../redux/reducers/projects/initialState';
+import initialExperimentState, { experimentTemplate } from '../../../redux/reducers/experiments/initialState';
 
-import { processUpload } from '../../utils/processUpload';
-import UploadStatus from '../../utils/UploadStatus';
+import { processUpload } from '../../../utils/upload/processUpload';
+import UploadStatus from '../../../utils/upload/UploadStatus';
 
 enableFetchMocks();
 
@@ -109,7 +109,7 @@ const flushPromises = () => new Promise(setImmediate);
 
 const mockStore = configureMockStore([thunk]);
 
-jest.mock('../../utils/loadAndCompressIfNecessary',
+jest.mock('../../../utils/upload/loadAndCompressIfNecessary',
   () => jest.fn().mockImplementation(
     (bundle) => {
       if (!bundle.valid) {
@@ -120,13 +120,13 @@ jest.mock('../../utils/loadAndCompressIfNecessary',
     },
   ));
 
-jest.mock('../../utils/environment', () => ({
+jest.mock('../../../utils/environment', () => ({
   __esModule: true,
   isBrowser: () => false,
   ssrGetCurrentEnvironment: () => 'development',
 }));
 
-jest.mock('../../redux/actions/samples/saveSamples', () => jest.fn().mockImplementation(() => ({
+jest.mock('../../../redux/actions/samples/saveSamples', () => jest.fn().mockImplementation(() => ({
   type: 'samples/saved',
 })));
