+++ src/components/data-management/FileUploadModal.jsx
@@ -16,9 +16,9 @@ import {
 } from 'antd';
 import { CheckCircleTwoTone, CloseCircleTwoTone, DeleteOutlined } from '@ant-design/icons';
 import Dropzone from 'react-dropzone';
-import techOptions from '../../utils/fileUploadSpecifications';
+import techOptions from '../../utils/upload/fileUploadSpecifications';
 import pushNotificationMessage from '../../utils/pushNotificationMessage';
-import { bundleToFile } from '../../utils/processUpload';
+import { bundleToFile } from '../../utils/upload/processUpload';
 
 const { Text, Title, Paragraph } = Typography;
 const { Option } = Select;
