+++ src/components/data-management/ProjectsListContainer.jsx
@@ -14,7 +14,7 @@ import FileUploadModal from './FileUploadModal';
 import { setActiveProject, updateProject, deleteProject as deleteProjectAction } from '../../redux/actions/projects';
 import PrettyTime from '../PrettyTime';
 
-import { processUpload } from '../../utils/processUpload';
+import { processUpload } from '../../utils/upload/processUpload';
 import validateInputs, { rules } from '../../utils/validateInputs';
 
 const ProjectsListContainer = (props) => {
