+++ src/utils/upload/loadAndCompressIfNecessary.js
